pyproject.toml
[project]
name = "swisslife"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "baml>=0.19.1",
    "baml-cli>=0.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.124.4",
    "httpx>=0.27.2",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.33.0",
]
